+++ .vscode/settings.json
@@ -3,6 +3,7 @@
     "editor.formatOnSave": true,
     "editor.formatOnPaste": true,
     "cSpell.words": [
+        "enduser",
         "favicon"
     ],
 }
