+++ server/db/json_schema/campaignSchema.js
@@ -1,0 +1,13 @@
+const campaignSchema = {
+  'type': 'object',
+  'properties': {
+    'userId': { 'type': 'string' },
+    'campaignId': { 'type': 'string' },
+    'points': { 'type': ['integer', 'null']},
+    'enduserId': { 'type': 'string' },
+    'partnerId': { 'type': 'string' }
+  },
+  'required': ['campaignId', 'enduserId', 'partnerId'],
+};
+
+module.exports = campaignSchema;
