+++ server/utils/mojaCampaigns/mojaCampaigns.js
@@ -1,0 +1,44 @@
+const CampaignMain = require('../../models/campaign_main');
+
+
+
+/**
+ *
+ * Handling redirect from moja campaign platform to
+ * sign up page with the following query params
+ *
+ *  @param {ctx} campaign_id
+ *  @param {ctx} points
+ *  @param {ctx} enduser_id
+ *  @param {ctx} partner_id
+ *  @param {*} userId
+ *
+ */
+
+async function mojaCampaigns(ctx, userId) {
+  let moja_query_params = {
+    userId: userId,
+    campaignId: ctx.request.body.user.campaign_id,
+    points: ctx.request.body.user.points || 1,
+    enduserId: ctx.request.body.user.enduser_id,
+    partnerId: ctx.request.body.user.partner_id
+  };
+
+  await CampaignMain.query.insertAndFetch(moja_query_params);
+}
+
+
+/**
+ * Linking Wikonnect account to moja account using a link button
+ * @param {*} ctx
+ * @param {*} next
+ */
+
+
+/**
+ TODO: Add fetch package for webhook request to the moja platform
+ * POST request accepting -> (campaign_id, points, enduser_id, partner_id, userId)
+ *
+ * */
+
+module.exports = { mojaCampaigns };
